Refactoring

+++ src/main/java/tasks/MaximumPerformanceOfTeam.java
@@ -16,13 +16,10 @@ import java.util.PriorityQueue;
 		url = "https://leetcode.com/problems/maximum-performance-of-a-team/",
 		difficulty = Difficulty.HARD
 )
-public class MaximumPerformanceOfTeam
-{
-	public int maxPerformance(int n, int[] speed, int[] efficiency, int k)
-	{
+public class MaximumPerformanceOfTeam {
+	public int maxPerformance(int n, int[] speed, int[] efficiency, int k) {
 		int[][] speedAndEfficiency = new int[n][2];
-		for (int i = 0; i < n; i++)
-		{
+		for (int i = 0; i < n; i++) {
 			speedAndEfficiency[i][0] = speed[i];
 			speedAndEfficiency[i][1] = efficiency[i];
 		}
@@ -31,17 +28,15 @@ public class MaximumPerformanceOfTeam
 		PriorityQueue<Integer> pq = new PriorityQueue<>();
 		long sum = 0;
 		long answer = 0;
-		for (int[] sAndE : speedAndEfficiency)
-		{
+		for (int[] sAndE : speedAndEfficiency) {
 			pq.add(sAndE[0]);
 
 			sum += sAndE[0];
-			if (pq.size() > k)
-			{
+			if (pq.size() > k) {
 				sum -= pq.remove();
 			}
 
-			answer = Math.max(answer, (long) sum * sAndE[1]);
+			answer = Math.max(answer, sum * sAndE[1]);
 		}
 
 		return (int) (answer % 1_000_000_007);
